src/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
dal-admin-filters==1.1.0
Django==4.1.2
django-autocomplete-light==3.9.4
django-bootstrap-form==3.4
django-filter==22.1
django_debug_toolbar==3.8.1
idna==3.4
mysqlclient==2.1.1
packaging==22.0
Pillow @ file:///C:/WorkSpace/packeteria/Pillow-9.3.0-cp310-cp310-win_amd64.whl
pip-review==1.3.0
pyparsing==3.0.9
python-decouple==3.6
reportlab @ file:///C:/WorkSpace/packeteria/reportlab-3.6.12-cp310-cp310-win_amd64.whl
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
